Add path visualization page
