--- source-code/safe/PathVisualization/templates/PathVisualization/nodatabase.html
@@ -1,71 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <title>Document</title>
-    <script src="https://a.amap.com/Loca/static/dist/jquery.min.js"></script>
-    <script src="https://webapi.amap.com/maps?key=2da678792428d85bd3f0c0e26b25c54b&v=1.4.15&plugin=Map3D"></script>
-    <script src="https://webapi.amap.com/loca?key=2da678792428d85bd3f0c0e26b25c54b&v=1.3.0"></script>
-</head>
-
-<body>
-    <div id='map' style='width:800px; height:600px;'></div>
-    <script>
-        window.onload = function() {
-            const map = new AMap.Map('map', {
-                pitch: 50, // 地图俯仰角度，有效范围 0 度- 83 度
-                viewMode: '3D', // 地图模式
-                mapStyle: 'amap://styles/8b1ff64d2c968f4f84c2a5f5ef562cfb', // 地图主题
-                zoom: 8, //设置地图的缩放级别
-                zooms: [8, 20], // 缩放级别限制
-                center: [116.368904, 39.913423],
-            });
-            AMap.plugin('AMap.DistrictSearch', function(){
-                var opts = {
-                    subdistrict: 0, // 获取边界不需要返回下级行政区
-                    extensions: "all", // 返回行政区边界坐标组等具体信息
-                    level: "city" // 查询行政级别为 市
-                };
-                var district = new AMap.DistrictSearch(opts);
-                district.search('北京市', function(status, result) {
-                    // 查询成功时，result即为对应的行政区信息
-                    var bounds = result.districtList[0].boundaries;
-
-                    for (var i = 0, l = bounds.length; i < l; i++) {
-                        var polygonbox = new AMap.Polygon({
-                            strokeWeight: 3,
-                            path: bounds[i],
-                            fillOpacity: 0.9,
-                            fillColor: "#6e3fc8",
-                            strokeColor: "#6e3fc8",
-                            extData: {
-                                IDname: "北京",
-                                Center: "传递内容"
-                            }
-                        });
-                        map.add(polygonbox);
-                    }
-                })
-                var lineArr=new Array();//创建线覆盖物节点坐标数组
-                lineArr.push(new AMap.LngLat("112.368904","39.913423"));
-                lineArr.push(new AMap.LngLat("116.382122","39.901176"));
-                lineArr.push(new AMap.LngLat("116.387271","39.912501"));
-                lineArr.push(new AMap.LngLat("118.398258","39.904600"));
-                polyline=new AMap.Polyline({
-                path:lineArr, //设置线覆盖物路径
-                strokeColor:"#3366FF", //线颜色
-                strokeOpacity:1, //线透明度
-                strokeWeight:5, //线宽
-                strokeStyle:"solid", //线样式
-                strokeDasharray:[10,5] //补充线样式
-                });
-                map.add(polyline);
-            })
-        }
-    </script>
-</body>
-
-</html>
