--- source-code/safe/PathVisualization/templates/PathVisualization/withdatabase.html
@@ -1,75 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
-    <title>Document</title>
-    <script src="https://a.amap.com/Loca/static/dist/jquery.min.js"></script>
-    <script src="https://webapi.amap.com/maps?key=2da678792428d85bd3f0c0e26b25c54b&v=1.4.15&plugin=Map3D"></script>
-    <script src="https://webapi.amap.com/loca?key=2da678792428d85bd3f0c0e26b25c54b&v=1.3.0"></script>
-</head>
-
-<body>
-    <div id='map' style='width:800px; height:600px;'></div>
-    <script>
-        window.onload = function() {
-            const map = new AMap.Map('map', {
-                pitch: 50, // 地图俯仰角度，有效范围 0 度- 83 度
-                viewMode: '3D', // 地图模式
-                mapStyle: 'amap://styles/57994c871bb604a4c79184f5f65d8782', // 地图主题
-                zoom: 8, //设置地图的缩放级别
-                zooms: [1, 20], // 缩放级别限制
-                center: [{{center.0}}, {{center.1}}], // 设置初始中心位置为快递最后更新的数据位置
-            });
-            AMap.plugin('AMap.DistrictSearch', function(){
-                var opts = {
-                    subdistrict: 0, // 获取边界不需要返回下级行政区
-                    extensions: "all", // 返回行政区边界坐标组等具体信息
-                    level: "province" // 查询行政级别为 市
-                };
-                var district = new AMap.DistrictSearch(opts);
-                {% for j in heatmapdata %}
-                    var province = '{{ j.0 }}';
-                    var color = '{{ j.1 }}';
-                    district.search(province, function(status, result) {
-                    // 查询成功时，result即为对应的行政区信息
-                    var bounds = result.districtList[0].boundaries;
-                    for (var i = 0, l = bounds.length; i < l; i++) {
-                        var polygonbox = new AMap.Polygon({
-                            strokeWeight: 1.5,
-                            path: bounds[i],
-                            fillOpacity: 0.8,
-                            fillColor: '{{ j.1 }}',
-                            strokeColor: "#87CEFA",
-                            extData: {
-                                IDname: "北京",
-                                Center: "传递内容"
-                            }
-                        });
-                        map.add(polygonbox);
-                    }
-                })
-                {% endfor %}
-
-                var lineArr=new Array();//创建线覆盖物节点坐标数组
-                {% for j in path_dot %}
-                    lineArr.push(new AMap.LngLat('{{ j.0 }}','{{ j.1 }}'))
-                {% endfor %}
-                polyline=new AMap.Polyline({
-                path:lineArr, //设置线覆盖物路径
-                strokeColor:"#FF0000", //线颜色
-                strokeOpacity:1, //线透明度
-                strokeWeight:5, //线宽
-                strokeStyle:"solid", //线样式
-                strokeDasharray:[10,5], //补充线样式
-                lineCap:"round" //线头部样式
-                });
-                map.add(polyline);
-            })
-        }
-    </script>
-</body>
-
-</html>
