--- source-code/safe/PathVisualization/urls.py
@@ -1,7 +1,0 @@
-from django.urls import path
-
-from . import views
-
-urlpatterns = [
-    path('<int:orderid>/', views.getdata, name='Visualization'),
-]
