+++ source-code/safe/dashboard/templates/information-summary-buyer.html
@@ -150,7 +150,7 @@ html,body,h1,h2,h3,h4,h5 {font-family: "Raleway", sans-serif}
 	<td>{{ product_order.onum }}</td>
 	<td>{{ product_order.sellername_id }}</td>
 	<td>{{ product_order.platformname_id }}</td>
-	<td><a href="#" %}" class="btn btn-sm btn-warning">查看</a></td>
+	<td><a href="{% url 'dashboard:visualization' delivery_order.dno %}" %}" class="btn btn-sm btn-warning">查看</a></td>
 	<td><a href="#" %}" class="btn btn-sm btn-warning">查看</a></td>
 </tr>
 {% endfor %}
